Move fetching burden into Trainers
