+++ src/components/Location/TrainerCard/TrainerCard.tsx
@@ -1,32 +1,16 @@
 import ItemDisplay from "@/components/Battle/ItemDisplay/ItemDisplay";
 import Battle from "@/models/Battle";
 import ItemData from "@/models/ItemData";
-import { fetchItems } from "@/utils/api";
 import Image from "next/image";
-import { useEffect, useState } from "react";
 import styles from "./TrainerCard.module.scss";
 
 type Props = {
     battle: Battle;
     game: string;
+    items: { [item: string]: ItemData };
 };
 
 const TrainerCard: React.FC<Props> = (props: Props) => {
-    // Fetched data
-    const [items, setItems] = useState<{ [item: string]: ItemData }>({});
-
-    // Fetch items on component load
-    useEffect(() => {
-        if (props.battle) {
-            const uniqueItems = [...new Set(props.battle.items)];
-            fetchItems(uniqueItems, props.game).then((items: ItemData[]) => {
-                const itemData: { [item: string]: ItemData } = {};
-                items.forEach((item: ItemData) => (itemData[item.slug] = item));
-                setItems(itemData);
-            });
-        }
-    }, [props.battle]);
-
     return (
         <div className={styles["trainer-card"]}>
             <div className={styles.header}>
@@ -44,11 +28,15 @@ const TrainerCard: React.FC<Props> = (props: Props) => {
                             {props.battle.trainer.class} {props.battle.name}
                         </p>
                         <p className={styles.location}>{props.battle.location}</p>
-                        {Object.keys(items).length > 0 ? (
+                        {props.battle.items.length > 0 ? (
                             <div className={styles.items}>
-                                {props.battle.items.map((item: string, key: number) => (
-                                    <ItemDisplay item={items[item]} showName={false} key={key} />
-                                ))}
+                                {props.battle.items.map((item: string, key: number) => {
+                                    return item in props.items ? (
+                                        <ItemDisplay item={props.items[item]} showName={false} key={key} />
+                                    ) : (
+                                        ""
+                                    );
+                                })}
                             </div>
                         ) : (
                             ""
